# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Architecture Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  role?: UserUpdateRole;
  /** @nullable */
  password?: string | null;
}